+++ api/v1/mdb/mongodb_types.go
@@ -8,6 +8,7 @@ import (
 
 	mdbcv1 "github.com/mongodb/mongodb-kubernetes-operator/api/v1"
 	"github.com/mongodb/mongodb-kubernetes-operator/pkg/kube/annotations"
+	"github.com/spf13/cast"
 
 	"github.com/10gen/ops-manager-kubernetes/pkg/util"
 	"github.com/10gen/ops-manager-kubernetes/pkg/util/env"
@@ -208,6 +209,7 @@ type DbSpec interface {
 	GetHorizonConfig() []MongoDBHorizonConfig
 	GetAdditionalMongodConfig() *AdditionalMongodConfig
 	GetExternalDomain() *string
+	GetMemberOptions() []MemberOptions
 }
 
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
@@ -239,6 +241,32 @@ type MongoDBConnectivity struct {
 	// +optional
 	ReplicaSetHorizons []MongoDBHorizonConfig `json:"replicaSetHorizons,omitempty"`
 }
+
+type MemberOptions struct {
+	Votes    *int              `json:"votes,omitempty"`
+	Priority *string           `json:"priority,omitempty"`
+	Tags     map[string]string `json:"tags,omitempty"`
+}
+
+func (o *MemberOptions) GetVotes() int {
+	if o.Votes != nil {
+		return cast.ToInt(o.Votes)
+	}
+	return 1
+}
+
+func (o *MemberOptions) GetPriority() float32 {
+	if o.Priority != nil {
+		return cast.ToFloat32(o.Priority)
+	}
+	return 1.0
+}
+
+func (o *MemberOptions) GetTags() map[string]string {
+	return o.Tags
+}
+
+// type MemberOptions map[string]interface{}
 
 type MongoDbStatus struct {
 	status.Common                   `json:",inline"`
@@ -313,6 +341,11 @@ type MongoDbSpec struct {
 	// DEPRECATED please use `spec.statefulSet.spec.serviceName` to provide a custom service name.
 	// this is an optional service, it will get the name "<rsName>-service" in case not provided
 	Service string `json:"service,omitempty"`
+
+	// MemberConfig
+	// +kubebuilder:pruning:PreserveUnknownFields
+	// +optional
+	MemberConfig []MemberOptions `json:"memberConfig,omitempty"`
 }
 
 func (s *MongoDbSpec) GetExternalDomain() *string {
@@ -324,6 +357,10 @@ func (s *MongoDbSpec) GetExternalDomain() *string {
 
 func (s *MongoDbSpec) GetHorizonConfig() []MongoDBHorizonConfig {
 	return s.Connectivity.ReplicaSetHorizons
+}
+
+func (s *MongoDbSpec) GetMemberOptions() []MemberOptions {
+	return s.MemberConfig
 }
 
 type SnapshotSchedule struct {