+++ controllers/om/replicaset/om_replicaset.go
@@ -25,7 +25,7 @@ func BuildFromStatefulSet(set appsv1.StatefulSet, dbSpec mdbv1.DbSpec) om.Replic
 func BuildFromStatefulSetWithReplicas(set appsv1.StatefulSet, dbSpec mdbv1.DbSpec, replicas int) om.ReplicaSetWithProcesses {
 	members := process.CreateMongodProcessesWithLimit(set, dbSpec, replicas)
 	replicaSet := om.NewReplicaSet(set.Name, dbSpec.GetMongoDBVersion())
-	rsWithProcesses := om.NewReplicaSetWithProcesses(replicaSet, members)
+	rsWithProcesses := om.NewReplicaSetWithProcesses(replicaSet, members, dbSpec.GetMemberOptions())
 	rsWithProcesses.SetHorizons(dbSpec.GetHorizonConfig())
 	return rsWithProcesses
 }
@@ -35,7 +35,7 @@ func BuildFromStatefulSetWithReplicas(set appsv1.StatefulSet, dbSpec mdbv1.DbSpe
 func BuildAppDBFromStatefulSet(set appsv1.StatefulSet, mdb omv1.AppDBSpec) om.ReplicaSetWithProcesses {
 	members := process.CreateAppDBProcesses(set, om.ProcessTypeMongod, mdb)
 	replicaSet := om.NewReplicaSet(set.Name, mdb.GetMongoDBVersion())
-	rsWithProcesses := om.NewReplicaSetWithProcesses(replicaSet, members)
+	rsWithProcesses := om.NewReplicaSetWithProcesses(replicaSet, members, mdb.GetMemberOptions())
 	return rsWithProcesses
 }